deep_research/local_deep_research/pyproject.toml
[project]
name = "local-deep-research"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agentscope[full]>=1.0.5",
    "argparse>=1.4.0",
    "fastapi>=0.115.12",
    "firecrawl-py>=2.5.3",
    "ibm-watsonx-ai>=1.3.0",
    "langchain-core>=1.0.2",
    "langchain-text-splitters>=0.3.8",
    "numpy>=1.26.4",
    "openai>=1.77.0",
    "pdfplumber>=0.11.6",
    "pydantic>=2.0.0",
    "pymilvus>=2.5.8",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "requests>=2.32.3",
    "termcolor>=3.1.0",
    "tqdm>=4.67.1",
    "uvicorn>=0.34.2",
]


[project.optional-dependencies]
all = [
    "unstructured-ingest>=1.0.24",
    "unstructured[all-docs]>=0.17.2",
    "docling>=2.15.1",
    "docling-core>=2.30.0",
    "crawl4ai>=0.6.2",
    "sentence-transformers>=4.1.0",
    "langchain-community>=0.3.2",
]